feat(back): refactor logic.js file to separate decision availability and processing logic and improve legibility
